Added toggle option
What I have done here is I have added a 3rd option that checks the state of the basestations, then toggles them to the opposite state. This is simply so you can run the script with toggle and not have to have multiple scripts to turn them on or off and just have one. The individual on and off are still left though in case you do not want to toggle an already off basestation on when trying to turn another off. Then again, it still takes into account the MAC(s) you provide so may not even need the on off part of it if you can just provide the mac and it toggles, then tells you what state it has been set to so you can confirm by looking at the Basestation's LED.